Pin X and/or Y axes
The pin buttons for the X and Y axes should be fully functional now. When an axis is pinned, recentering will not affect it, nor can the axis be changed by the coord control widget.
I also got the recenter button working, but that was a quick fix. It is very useful in conjunction with the pin buttons.