Zoom in and out, Better CoordControl sync
I worked on several different features. First off, I worked on getting CoordControl better connected with the PolygonView such that changing the perspective automatically changes the text for the CoordControl.
Secondly, I got zoom in and zoom out working, making zoom out increase the perspective 150% and zoom in increasing 66% (which is the reverse of increasing 150%).