Perimeter and Work on README
I added a new perimeter field to the shape that will update whenever the perimeter of the shape changes. It is a static text field, so the user can see the data, copy it, but not alter it.

I also started working on the read me file. It shouldn't be too long, but I want to make sure that the screen shot image I have is working properly.