Caption buttons
I added captions that pop up if you hover over buttons for half a second. Because of redirection of pixel location, it was not possible to do this without some hardcoded values.

Also, the whole procedure will cause concurrent access of the component list, which causes a warning error at runtime.